ExploreSection extraction, small tweaks and fixes
